gagal reset & settings

+++ WORDLE_SEM2/main.py
@@ -1,6 +1,5 @@
 import pygame
 import os
-import pygame.mixer
 import moviepy.editor as mp
 import random
 import sys
@@ -20,8 +19,7 @@ OUTLINE = "#d3d6da"
 FILLED_OUTLINE = "#878a8c"
 CORRECT_WORD = "ajaib"
 ALPHABET = ["ABCDEFGHIJKLM", "NOPQRSTUVWXYZ"]
-GUESSED_LETTER_FONT_PATH = "/Users/jennifernathaniahartono/Documents/WORDLE_PROJ_SEM2/WORDLE_SEM2/assets/FreeSansBold.otf"
-#GUESSED_LETTER_FONT_PATH = "C:\\Users\\Angeline\\Documents\\GitHub\\WORDLE_PROJ_SEM2\\WORDLE_SEM2\\assets\\FreeSansBold.otf"
+GUESSED_LETTER_FONT_PATH = "/Users/jennifernathaniahartono/Documents/WORDLE_PROJ_SEM2-5/WORDLE_SEM2/assets/sfx/TypoSlab.otf"
 AVAILABLE_LETTER_FONT_PATH = GUESSED_LETTER_FONT_PATH
 
 # Setup display
@@ -29,33 +27,41 @@ SCREEN = pygame.display.set_mode((WIDTH, HEIGHT))
 pygame.display.set_caption("NYASAR KATA GAME")
 
 # Load assets
-#jenni
-video_path = "/Users/jennifernathaniahartono/Documents/WORDLE_PROJ_SEM2-2/WORDLE_SEM2/assets/bg_cth.mp4"
-video_path_2 = "/Users/jennifernathaniahartono/Documents/WORDLE_PROJ_SEM2-2/WORDLE_SEM2/assets/bg_vid.mp4"
+# jenni
+video_path = "/Users/jennifernathaniahartono/Documents/WORDLE_PROJ_SEM2-5/WORDLE_SEM2/assets/bg_ingame.mp4"
+video_path_2 = "/Users/jennifernathaniahartono/Documents/WORDLE_PROJ_SEM2-5/WORDLE_SEM2/assets/bg_vid.mp4"
 video = mp.VideoFileClip(video_path)
 video2 = mp.VideoFileClip(video_path_2)
 
-#angel
-#video_path = "C:\\Users\\Angeline\\Documents\\GitHub\\WORDLE_PROJ_SEM2\\WORDLE_SEM2\\assets\\bg_cth.mp4"
-#video_path_2 = "C:\\Users\\Angeline\\Documents\\GitHub\\WORDLE_PROJ_SEM2\\WORDLE_SEM2\\assets\\bg_vid.mp4"
-#video = mp.VideoFileClip(video_path)
-#video2 = mp.VideoFileClip(video_path_2)
+#winScene
+win_video = mp.VideoFileClip("/Users/jennifernathaniahartono/Documents/WORDLE_PROJ_SEM2-5/WORDLE_SEM2/assets/win_video.mp4")
+lose_video = mp.VideoFileClip("/Users/jennifernathaniahartono/Documents/WORDLE_PROJ_SEM2-5/WORDLE_SEM2/assets/lose_video.mp4")
+
+# angel
+# video_path = "C:\\Users\\Angeline\\Documents\\GitHub\\WORDLE_PROJ_SEM2\\WORDLE_SEM2\\assets\\bg_cth.mp4"
+# video_path_2 = "C:\\Users\\Angeline\\Documents\\GitHub\\WORDLE_PROJ_SEM2\\WORDLE_SEM2\\assets\\bg_vid.mp4"
+# video = mp.VideoFileClip(video_path)
+# video2 = mp.VideoFileClip(video_path_2)
 
 # Extract audio from the main video
-audio_path = "temp_audio.mp3"
-video.audio.write_audiofile(audio_path)
+#audio_path = "/Users/jennifernathaniahartono/Documents/WORDLE_PROJ_SEM2-4/WORDLE_SEM2/assets/sfx/gamelan.mp3"
+#audio = pygame.mixer.music.load(audio_path)
+audio_sound = pygame.mixer.Sound("/Users/jennifernathaniahartono/Documents/WORDLE_PROJ_SEM2-5/WORDLE_SEM2/assets/sfx/gamelan.mp3")
 
-# Load audio
-pygame.mixer.music.load(audio_path)
+# # Directly load the MP3 if available
+# try:
+#     if not os.path.exists(audio_path):
+#         # Extract audio only if the mp3 doesn't exist
+#         video.audio.write_audiofile(audio_path)
+# except Exception as e:
+#     print(f"Error extracting audio: {e}")
+
 
 # Load other assets
-#ICON = pygame.image.load("C:\\Users\\Angeline\\Documents\\GitHub\\WORDLE_PROJ_SEM2\\WORDLE_SEM2\\assets\\tiles.png")
 ICON = pygame.image.load("/Users/jennifernathaniahartono/Documents/WORDLE_PROJ_SEM2/WORDLE_SEM2/assets/tiles.png")
 
 pygame.display.set_icon(ICON)
 
-#ENTER_SOUND = pygame.mixer.Sound("C:\\Users\\Angeline\\Documents\\GitHub\\WORDLE_PROJ_SEM2\\WORDLE_SEM2\\assets\\sfx\\button_1.ogg")
-#TYPE_SOUND = pygame.mixer.Sound("C:\\Users\\Angeline\\Documents\\GitHub\\WORDLE_PROJ_SEM2\\WORDLE_SEM2\\assets\\sfx\\type.ogg")
 ENTER_SOUND = pygame.mixer.Sound("/Users/jennifernathaniahartono/Documents/GitHub/WORDLE_PROJ_SEM2/WORDLE_SEM2/assets/sfx/button_1.ogg")
 TYPE_SOUND = pygame.mixer.Sound("/Users/jennifernathaniahartono/Documents/GitHub/WORDLE_PROJ_SEM2/WORDLE_SEM2/assets/sfx/type.ogg")
 
@@ -74,30 +80,25 @@ current_scene = "opening"  # New variable to track the current scene
 video_start_time = pygame.time.get_ticks()  # Initialize video_start_time
 
 class Button:
-    def __init__(self, text, position, width, height, font):
+    def __init__(self, text, position, font):
         self.text = text
         self.position = position
-        self.width = width
-        self.height = height
         self.font = font
-        self.rect = pygame.Rect(position[0], position[1], width, height)
-        self.color = GREY
-
+        self.text_surface = self.font.render(self.text, True, "#516061")
+        self.rect = self.text_surface.get_rect(center=self.position)
+        
     def draw(self):
-        pygame.draw.rect(SCREEN, self.color, self.rect)
-        text_surface = self.font.render(self.text, True, "white")
-        text_rect = text_surface.get_rect(center=self.rect.center)
-        SCREEN.blit(text_surface, text_rect)
+        audio_sound.play()
+        SCREEN.blit(self.text_surface, self.rect)
 
     def is_clicked(self, pos):
         return self.rect.collidepoint(pos)
 
 class OpeningScene:
     def __init__(self):
-        self.title_font = pygame.font.Font(GUESSED_LETTER_FONT_PATH, 75)
-        self.button_font = pygame.font.Font(GUESSED_LETTER_FONT_PATH, 50)
-        self.start_button = Button("Start", (225, 300), 200, 75, self.button_font)
-        self.settings_button = Button("Settings", (225, 400), 200, 75, self.button_font)
+        self.button_font = pygame.font.Font(GUESSED_LETTER_FONT_PATH, 30)
+        self.start_button = Button(   "S T A R T   ", (335, 441), self.button_font)
+        self.settings_button = Button("S E T T I N G S", (325, 551), self.button_font)
         self.video_start_time = pygame.time.get_ticks()  # Initialize local video_start_time
         
     def draw(self):
@@ -108,19 +109,18 @@ class OpeningScene:
             SCREEN.blit(frame_surface, (0, 0))
         else:
             self.video_start_time = pygame.time.get_ticks()
-        
-        # title_surface = self.title_font.render("Wordle Game", True, "white")
-        # title_rect = title_surface.get_rect(center=(WIDTH / 2, 150))
-        # SCREEN.blit(title_surface, title_rect)
+        #pygame.mixer.music.play(-1)
         self.start_button.draw()
         self.settings_button.draw()
 
     def handle_event(self, event):
         global current_scene, video_start_time
+        #pygame.mixer.music.load(audio_path)
         if event.type == pygame.MOUSEBUTTONDOWN:
             if self.start_button.is_clicked(event.pos):
                 current_scene = "gameplay"
                 video_start_time = pygame.time.get_ticks()  # Update global video_start_time
+                audio_sound.stop()
             elif self.settings_button.is_clicked(event.pos):
                 current_scene = "settings"
                 video_start_time = pygame.time.get_ticks()  # Update global video_start_time
@@ -129,7 +129,7 @@ class SettingsScene:
     def __init__(self):
         self.title_font = pygame.font.Font(GUESSED_LETTER_FONT_PATH, 75)
         self.button_font = pygame.font.Font(GUESSED_LETTER_FONT_PATH, 50)
-        self.back_button = Button("Back", (225, 500), 200, 75, self.button_font)
+        self.back_button = Button("Back", (225, 500), self.button_font)
         self.video_start_time = pygame.time.get_ticks()  # Initialize local video_start_time
 
     def draw(self):
@@ -195,6 +195,7 @@ for row in ALPHABET:
         indicator_x += 50
     indicator_y += 50
 
+
 def check_guess(guess_to_check):
     global current_guess, current_guess_string, guesses_count, current_letter_bg_x, game_result
     game_decided = False
@@ -239,15 +240,77 @@ def check_guess(guess_to_check):
     if guesses_count == 5 and game_result == "":
         game_result = "L"
 
-def play_again():
-    pygame.draw.rect(SCREEN, "grey", (10, 600, 1000, 600))
-    play_again_font = pygame.font.Font(GUESSED_LETTER_FONT_PATH, 50)
-    play_again_text = play_again_font.render("Press ENTER to Play Again!", True, "black")
-    play_again_rect = play_again_text.get_rect(center=(WIDTH / 2, 700))
-    word_was_text = play_again_font.render(f"The word was {CORRECT_WORD}!", True, "black")
-    word_was_rect = word_was_text.get_rect(center=(WIDTH / 2, 650))
-    SCREEN.blit(word_was_text, word_was_rect)
-    SCREEN.blit(play_again_text, play_again_rect)
+# def play_again():
+#     pygame.draw.rect(SCREEN, "grey", (10, 600, 1000, 600))
+#     play_again_font = pygame.font.Font(GUESSED_LETTER_FONT_PATH, 50)
+#     play_again_text = play_again_font.render("Press ENTER to Play Again!", True, "black")
+#     play_again_rect = play_again_text.get_rect(center=(WIDTH / 2, 700))
+#     word_was_text = play_again_font.render(f"The word was {CORRECT_WORD}!", True, "black")
+#     word_was_rect = word_was_text.get_rect(center=(WIDTH / 2, 650))
+#     SCREEN.blit(word_was_text, word_was_rect)
+#     SCREEN.blit(play_again_text, play_again_rect)
+
+class WinScene:
+    def __init__(self):
+        self.button_font = pygame.font.Font(GUESSED_LETTER_FONT_PATH, 30)
+        self.play_again_button = Button(   " P L A Y  G A M E ", (335, 386), self.button_font)
+        self.main_menu_button = Button("   M A I N  M E N U ", (325, 505), self.button_font)
+        self.video_start_time = pygame.time.get_ticks()  # Initialize local video_start_time
+        
+    def draw(self):
+        current_time = (pygame.time.get_ticks() - self.video_start_time) / 1000.0
+        if current_time < win_video.duration:
+            frame = win_video.get_frame(current_time)
+            frame_surface = pygame.surfarray.make_surface(frame.swapaxes(0, 1))
+            SCREEN.blit(frame_surface, (0, 0))
+        else:
+            self.video_start_time = pygame.time.get_ticks()
+        #pygame.mixer.music.play(-1)
+        self.play_again_button.draw()
+        self.main_menu_button.draw()
+
+    def handle_event(self, event):
+        global current_scene, video_start_time
+        #pygame.mixer.music.load(audio_path)
+        if event.type == pygame.MOUSEBUTTONDOWN:
+            if self.play_again_button.is_clicked(event.pos):
+                current_scene = "gameplay"
+                video_start_time = pygame.time.get_ticks()  # Update global video_start_time
+                audio_sound.stop()
+            elif self.main_menu_button.is_clicked(event.pos):
+                current_scene = "opening"
+                video_start_time = pygame.time.get_ticks()  # Update global video_start_time
+                
+class LoseScene:
+    def __init__(self):
+        self.button_font = pygame.font.Font(GUESSED_LETTER_FONT_PATH, 30)
+        self.play_again_button = Button(   "P L A Y  G A M E ", (335, 425), self.button_font)
+        self.main_menu_button = Button("M A I N  M E N U ", (325, 540), self.button_font)
+        self.video_start_time = pygame.time.get_ticks()  # Initialize local video_start_time
+        
+    def draw(self):
+        current_time = (pygame.time.get_ticks() - self.video_start_time) / 1000.0
+        if current_time < lose_video.duration:
+            frame = lose_video.get_frame(current_time)
+            frame_surface = pygame.surfarray.make_surface(frame.swapaxes(0, 1))
+            SCREEN.blit(frame_surface, (0, 0))
+        else:
+            self.video_start_time = pygame.time.get_ticks()
+        #pygame.mixer.music.play(-1)
+        self.play_again_button.draw()
+        self.main_menu_button.draw()
+
+    def handle_event(self, event):
+        global current_scene, video_start_time
+        #pygame.mixer.music.load(audio_path)
+        if event.type == pygame.MOUSEBUTTONDOWN:
+            if self.play_again_button.is_clicked(event.pos):
+                current_scene = "gameplay"
+                video_start_time = pygame.time.get_ticks()  # Update global video_start_time
+                audio_sound.stop()
+            elif self.main_menu_button.is_clicked(event.pos):
+                current_scene = "opening"
+                video_start_time = pygame.time.get_ticks()  # Update global video_start_time
 
 def reset():
     global guesses_count, CORRECT_WORD, guesses, current_guess, current_guess_string, game_result
@@ -291,10 +354,13 @@ fps = video.fps
 
 opening_scene = OpeningScene()
 settings_scene = SettingsScene()
+win_scene = WinScene()
+lose_scene = LoseScene()
 
 while True:
     if current_scene == "gameplay":
         current_time = (pygame.time.get_ticks() - video_start_time) / 1000.0
+        
         if current_time < video.duration:
             frame = video.get_frame(current_time)
             frame_surface = pygame.surfarray.make_surface(frame.swapaxes(0, 1))
@@ -311,6 +377,7 @@ while True:
 
         if current_scene == "opening":
             opening_scene.handle_event(event)
+        
         elif current_scene == "settings":
             settings_scene.handle_event(event)
 
@@ -323,17 +390,25 @@ while True:
                         if len(current_guess_string) == 5 and current_guess_string.lower() in WORDS:
                             check_guess(current_guess)
                             ENTER_SOUND.play()
-                            if game_result != "":
-                                play_again()
+                            if game_result == "W":
+                                current_scene = "win"
+                            elif game_result == "L":
+                                current_scene = "lose"
                 elif event.key == pygame.K_BACKSPACE:
-                    if len(current_guess_string) > 0:
+                    if len(current_guess_string) > 0 and game_result == "":
                         delete_letter()
                 else:
                     key_pressed = event.unicode.upper()
-                    if key_pressed in "QWERTYUIOPASDFGHJKLZXCVBNM":
+                    if key_pressed in "QWERTYUIOPASDFGHJKLZXCVBNM" and game_result == "":
                         if len(current_guess_string) < 5:
                             create_new_letter(key_pressed)
                             TYPE_SOUND.play()
+                            
+        if current_scene == "win":
+            win_scene.handle_event(event)
+            
+        elif current_scene == "lose":
+            lose_scene.handle_event(event)
 
     if current_scene == "gameplay":
         for guess in guesses:
@@ -345,21 +420,25 @@ while True:
 
         # Start the video and audio when entering the gameplay scene
         if not pygame.mixer.music.get_busy():
-            pygame.mixer.music.play(-1)
+            #pygame.mixer.music.play(-1)
             video_start_time = pygame.time.get_ticks()
 
     else:
         # Stop the video and audio when leaving the gameplay scene
         if pygame.mixer.music.get_busy():
             pygame.mixer.music.stop()
-
         if current_scene == "opening":
             opening_scene.draw()
         elif current_scene == "settings":
             settings_scene.draw()
+        elif current_scene == "win":
+            win_scene.draw()
+        elif current_scene == "lose":
+            lose_scene.draw()
 
     pygame.display.update()
     clock.tick(fps)
 
-# Clean up temporary files
-os.remove(audio_path)
+# Clean up temporary files if they were created
+if not os.path.exists(audio_path):
+    os.remove(audio_path)
